ai_assignment_checking/requirements.txt
groq
scikit-learn
python-dotenv
